closes GaelGoncalba#7, elección del lenguaje
